WIP categoryController

+++ src/main/java/com/example/labb2/service/interfaces/ICategoryService.java
@@ -4,6 +4,6 @@ import com.example.labb2.dto.model.CategoryDto;
 
 public interface ICategoryService {
     CategoryDto getAllCategories();
-    CategoryDto getCategoryById(int categoryId);
+    CategoryDto getCategoryById(long categoryId);
     void createCategory(String name, String symbol, String description);
 }
